bpo-32075: Expose ZipImporter Type Object in the include header files.
This should hopefully allow subclassing the zipimporter class outside of
the Python layer as well.
